add 2 compute plan tests: success + recursive fail
